// isoft-fe/src/app/models/salary-socail.model.ts
export interface SalarySocial {
    id: number;
    code: string;
    name: string;
    accountDebit: string;
    detailDebit1: string;
    detailDebit2: string;
    accountCredit: string;
    detailCredit1: string;
    detailCredit2: string;
    order: number;
    valueCompany: any;
    valueUser: any;
}